requirements.txt
amqp==1.4.6
anyjson==0.3.3
astroid==1.3.8
autopep8==1.2.1
billiard==3.3.0.20
celery==3.1.18
celery-haystack==0.9
certifi==2015.9.6.2
Django==1.8.4
django-activelink==0.4
django-admin-bootstrapped==2.5.6
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-cacheback==0.9.1
django-celery==3.1.16
django-celery-transactions==0.3.2
django-crispy-forms==1.5.2
django-debug-toolbar==1.3.2
django-haystack==2.4.0
django-redis-cache==1.6.0
django-rest-swagger==0.3.4
djangorestframework==3.2.3
elasticsearch==1.6.0
factory-boy==2.5.2
freeze==1.0.9
kombu==3.0.26
logilab-common==1.0.2
pbr==1.8.0
pep8==1.6.2
pyelasticsearch==1.4
pylint==1.4.4
python-memcached==1.57
pytz==2015.4
PyYAML==3.11
redis==2.10.3
simplejson==3.8.0
six==1.9.0
sqlparse==0.1.16
stevedore==1.8.0
urllib3==1.12
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
Whoosh==2.7.0
